Windows port resolution:
- .gitattributes: keep both (fixture eol=lf + spdlog -whitespace)
- src/CMakeLists.txt: keep both (ninfer_product_logging link + WIN32 bcrypt)
- load_progress.cpp / console_log.cpp: accept upstream deletion
  (replaced by unified product logging; localtime_s port fix no longer needed,
  new logging code is cross-platform)
- tests: take upstream (BUILD_TESTING defaults OFF on this tree)
Port fixes carried forward: gmtime_s (38d741a3), bcrypt+_WIN32 (71f0c903),
NVTX_DISABLE (cc38a826); reader.cpp CreateFileW branch untouched by upstream

+++ bench/ops/sliding_window_attention_bench.cu
@@ -162,7 +162,7 @@ Options parse_options(int argc, char** argv) {
 CyclicKVCacheLayerView make_context_view(DeviceBuffer& k, DeviceBuffer& v) {
     return {
         .k               = Tensor(k.p, DType::BF16, {kHeadDim, kWindow, kKvHeads, 1}),
-        .v               = Tensor(v.p, DType::BF16, {kHeadDim, kWindow, kKvHeads, 1}),
+        .v               = Tensor(v.p, DType::FP16, {kHeadDim, kWindow, kKvHeads, 1}),
         .capacity        = kWindow,
         .padded_capacity = kWindow,
         .num_kv_heads    = kKvHeads,
